Add a simple conveyor belt
Haven't yet had a chance to test if it works with a person, but I've added a simple conveyor belt to the scene that can move objects that have been placed on it.
However, I'm not sure if any of the cameras and aliases attached to the player will be detected as "colliding" with the conveyor belt -- do we need to make the camerarig a child of some physical avatar?